把github_helper相关的操作放到service层.
