+++ src/controllers/authorize.go
@@ -2,6 +2,7 @@ package controllers
 
 import (
 	"Kcoin-Golang/src/models"
+	"Kcoin-Golang/src/service"
 	"fmt"
 	"github.com/astaxie/beego"
 	"github.com/astaxie/beego/orm"
@@ -11,40 +12,40 @@ import (
 type AuthoController struct {
 	beego.Controller
 }
+
 var currentUserId string
+
 func (c *AuthoController) Get() {
 	var code string = c.GetString("code")
-	accessToken, _ := getAccessToken(code)
-	text := getUserJson(accessToken)
+	accessToken, _ := service.GetAccessToken(code)
+	text := service.GetUserJson(accessToken)
 
 	name := text.Data.Name
 	id := text.Data.Id
-	currentUserId=id
+	currentUserId = id
 	// 修改参数
-	GithubUser.setGithubUserAccessToken(id,name, accessToken)
+	service.GithubUser.SetGithubUserAccessToken(id, name, accessToken)
 	uri := text.Data.Uri
 
 	o := orm.NewOrm()
 	o.Using("default")
 	//  移到model 改成GitID查询
-	res , _ := models.FinduserByGitId(id)
-	
+	res, _ := models.FinduserByGitId(id)
+
 	if res.UserId == "" {
-		err := models.InsertUser(name,uri,id)
+		err := models.InsertUser(name, uri, id)
 
 		if err != nil {
 			panic(err)
 		}
-	}else{
+	} else {
 		time := time.Now().Format("2006-01-02 15:04:05.000000")
 		updateSql := `update "K_User" set register_time = ? where GITHUB_USER_ID = ?`
-		_,err := o.Raw(updateSql, time, id).Exec()
+		_, err := o.Raw(updateSql, time, id).Exec()
 		if err != nil {
 			panic(err)
 		}
 	}
-
-
 
 	//存储用户名到cooike中，获取语法：c.Ctx.GetCookie("userName")
 	c.Ctx.SetCookie("userName", text.Data.Name, 3600)
@@ -53,7 +54,7 @@ func (c *AuthoController) Get() {
 	//存储用户头像url到cooike中，获取语法：c.Ctx.GetCookie("headShotUrl")
 	c.Ctx.SetCookie("headShotUrl", text.Data.Uri, 3600)
 	//存储用户登录状态到cooike中，其中1表示已登录，获取语法：c.Ctx.GetCookie("status")
-  
+
 	c.Ctx.SetCookie("status", string('1'), 3600)
 
 	if redirectUrl := c.Ctx.GetCookie("lastUri"); redirectUrl != "" {
