+++ src/service/github_helper.go
@@ -1,15 +1,18 @@
 package service
 
 import (
+	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/astaxie/beego"
 	"io/ioutil"
 	"net/http"
 	"strings"
 	"time"
 )
 
-//解析向api:"https://api.github.com/repos/"+userName+"/"+programName发送请求后的返回值
+// struct for github api: "https://api.github.com/repos/"+userName+"/"+programName发送请求后的返回值
 type JsonData struct {
 	ID       int    `json:"id"`
 	NodeID   string `json:"node_id"`
@@ -127,6 +130,7 @@ type JsonData struct {
 	SubscribersCount int `json:"subscribers_count"`
 }
 
+// struct for github api: "https://api.github.com/repos/" + userName + "/" + programName + "/" + "contributors"
 type ContributorData []struct {
 	Login             string `json:"login"`
 	ID                int    `json:"id"`
@@ -149,21 +153,67 @@ type ContributorData []struct {
 	Contributions     int    `json:"contributions"`
 }
 
-/*
-
-获取项目star数量的接口
-函数名：GetStarNum
-函数参数：url string
-返回值：starNum int 返回url对应项目的star数目
-
-使用实例：
-func main() {
-	var info =getStarNum("https://github.com/OS-ABC/Kcoin-Golang")
-	fmt.Println(info)
+// struct for github api: "https://api.github.com/users/" + username
+type API_User struct {
+	Login               string `json:"login"`
+	Id                  int    `json:"id"`
+	Node_id             string `json:"node_id"`
+	Avatar_url          string `json:"avatar_url"`
+	Gravatar_id         string `json:"gravatar_id"`
+	Url                 string `json:"url"`
+	Html_url            string `json:"html_url"`
+	Followers_url       string `json:"followers_url"`
+	Following_url       string `json:"following_url"`
+	Gists_url           string `json:"gists_url"`
+	Starred_url         string `json:"starred_url"`
+	Subscriptions_url   string `json:"subscriptions_url"`
+	Organizations_url   string `json:"organizations_url"`
+	Repos_url           string `json:"repos_url"`
+	Events_url          string `json:"events_url"`
+	Received_events_url string `json:"received_events_url"`
+	Type                string `json:"type"`
+	Site_admin          bool   `json:"site_admin"`
+	Name                string `json:"name"`
+	Company             string `json:"company"`
+	Blog                string `json:"blog"`
+	Location            string `json:"location"`
+	Email               string `json:"email"`
+	Hireable            bool   `json:"hireable"`
+	Bio                 string `json:"bio"`
+	Public_repos        int    `json:"public_repos"`
+	Public_gists        int    `json:"public_gists"`
+	Followers           int    `json:"followers"`
+	Following           int    `json:"following"`
+	Created_at          string `json:"created_at"`
+	Updated_at          string `json:"updated_at"`
 }
 
-*/
+/**
+ * 这是一个全局数据结构，目前只有三个字段，用来保存GithubID, Github Name和对应access_token
+ */
+type GithubInfo struct {
+	GithubId    string
+	GithubName  string
+	AccessToken string
+}
 
+// TODO 妥善使用该数据结构, 用户信息应该用session保存, 可以建立一个sessionID->session的映射, 但是不太清楚session如何使用, 这里需要会的人来修改这个丑陋的数据结构
+//Github UserID -> GithubInfo
+type GithubUserMap map[string]*GithubInfo
+
+var GithubUser GithubUserMap
+
+func init() {
+	fmt.Println("Controller initialized!")
+	GithubUser = make(GithubUserMap)
+}
+
+/**
+ * 获取项目star数量的接口
+ * 函数名：GetStarNum
+ * 函数参数：url string
+ * 返回值：starNum int 返回url对应项目的star数目
+ */
 func GetStarNum(url string) int {
 	var starNum = 0
 	//从url中获取到用户名和项目名
@@ -224,4 +274,151 @@ func GetContributorNum(url string) int {
 	str_arr := strings.Split(res, " ")
 	count := len(str_arr)
 	return count
+}
+
+type Data struct {
+	Id   string `json:userId`
+	Name string `json:"userName"`
+	Uri  string `json:"headShotUrl"`
+}
+
+type UserJson struct {
+	ErrorCode int
+	Data      Data
+}
+
+func GetUserJson(access_token string) UserJson {
+	var url_2 = "https://api.github.com/user?" + "access_token=" + access_token
+
+	client_2 := &http.Client{}
+	response_2, _ := client_2.Get(url_2)
+	defer response_2.Body.Close()
+	body_2, err_2 := ioutil.ReadAll(response_2.Body)
+	if err_2 != nil {
+		panic(err_2)
+	}
+
+	// 获取ID
+	var name = strings.Split(strings.Split(string(body_2), ",")[0], "\"")[3]
+	var uri = strings.Split(strings.Split(string(body_2), ",")[3], "\"")[3]
+	var id = strings.Split(strings.Split(string(body_2), ",")[1], ":")[1]
+
+	//select id according to name
+	data := Data{
+		Name: name,
+		Uri:  uri,
+		Id:   id,
+	}
+	userJson := UserJson{
+		ErrorCode: 0,
+		Data:      data,
+	}
+
+	return userJson
+}
+
+func GetAccessToken(code string) (accessToken string, err error) {
+	client_id := beego.AppConfig.String("client_id")
+	client_secret := beego.AppConfig.String("client_secret")
+	url_1 := "https://github.com/login/oauth/access_token?code=" + code + "&client_id=" + client_id + "&client_secret=" + client_secret
+
+	client := &http.Client{}
+	response, err := client.Get(url_1)
+	if err != nil {
+		return "", err
+	} else {
+		defer response.Body.Close()
+	}
+	body, err1 := ioutil.ReadAll(response.Body)
+	if err1 != nil {
+		panic(err1)
+	}
+	accessToken = strings.Split(strings.Split(string(body), "&")[0], "=")[1]
+	return accessToken, err
+}
+
+/**
+ * 设置Github User这个map的Access Token字段.
+ */
+func (this GithubUserMap) SetGithubUserAccessToken(id string, name string, accessToken string) {
+	if _, ok := this[id]; !ok {
+		this[id] = new(GithubInfo)
+	}
+	this[id].AccessToken = accessToken
+	this[id].GithubId = id
+	this[id].GithubName = name
+}
+
+func (this GithubUserMap) GetGithubUserAccessToken(userId string) (string, error) {
+	if userInfo, ok := this[userId]; ok {
+		return userInfo.AccessToken, nil
+	} else {
+		return "", fmt.Errorf("user id %s is not valid", userId)
+	}
+}
+
+//getWebhooksUrl 可以通过
+func RegisterGithubWebhooks(userId string, repoName string) {
+	accessToken, _ := GithubUser.GetGithubUserAccessToken(userId)
+	postPayload := getPayloadOfRegisterGithubWebhooks()
+	userName := GithubUser[userId].GithubName
+	api_url := getWebhooksUrlBy(userName, repoName)
+	bytePostPayload := []byte(postPayload)
+	buffer := bytes.NewBuffer(bytePostPayload)
+	request, err := http.NewRequest("POST", api_url, buffer)
+	if err != nil {
+		fmt.Printf("http.NewRequest%v", err)
+	}
+	request.Header.Set("Content-Type", "application/json")
+	request.Header.Set("Authorization", "token "+accessToken)
+	client := &http.Client{}
+	resp, err := client.Do(request.WithContext(context.TODO()))
+	if err != nil {
+		fmt.Printf("client.Do%v", err)
+	}
+	respBytes, err := ioutil.ReadAll(resp.Body)
+	defer resp.Body.Close()
+	if err != nil {
+		fmt.Printf("ioutil.ReadAll%v", err)
+	}
+
+	fmt.Println(string(respBytes))
+}
+
+/**
+ * TODO 在config中设置secret, 与github_webhooks中的Post函数一起完成
+ */
+func getPayloadOfRegisterGithubWebhooks() string {
+	return `{
+  "name": "web",
+  "active": true,
+  "events": [
+    "push",
+    "pull_request"
+  ],
+  "config": {
+    "url": "http://114.115.206.8:8080/webhooks",
+    "content_type": "json",
+    "insecure_ssl": "0"
+  }
+}`
+}
+
+func getWebhooksUrlBy(userName string, repoName string) string {
+	return "https://api.github.com/repos/" + userName + "/" + repoName + "/hooks"
+}
+
+func GetGithubId(username string) int {
+	api := "https://api.github.com/users/" + username
+	client := &http.Client{}
+	response, _ := client.Get(api)
+	defer response.Body.Close()
+	body, err_1 := ioutil.ReadAll(response.Body)
+	if err_1 != nil {
+		panic(err_1)
+	}
+	var res API_User
+	json.Unmarshal(body, &res)
+	fmt.Println(res.Id)
+	return res.Id
 }
